Catcher implemented rules
I have implemented 15 rules for the catcher, you can see the description
of each rule on the captureRulesKB file. I have added 3(i think)
functions on the knwoledge.py which have to do with the flanking part,
they are weird implementation wise and if you don't understand from the
comments feel free to ask me whenever you want, maybe there are some
bugs and shit :P. I have tested the rules but not on the dynamicscipting
file.